Move completed, references updated, minor error fix to ICourseLoader typ management
