+++ src/features/admin/courses/loaders/courseLoader.ts
@@ -2,14 +2,15 @@ import { LoaderFunctionArgs } from "react-router";
 import { CustomError } from "../../../shared/classes";
 import { BASE_URL } from "../../../shared/constants";
 import { fetchWithToken } from "../../../shared/utilities";
-import { ICourse, ICourseLoader } from "../../types";
+import { ICourse, ICourseLoader } from "../types";
 
 export const courseLoader = async ({ params }: LoaderFunctionArgs): Promise<ICourseLoader> => {
   if (!params.id) throw new Response("Missing id", { status: 400 });
 
   try {
-    const data = fetchWithToken<ICourse>(`${BASE_URL}/courses/${params.id}`);
-    return { course: data };
+    const data = await fetchWithToken<ICourse>(`${BASE_URL}/courses/${params.id}`);
+    if (!data) throw new Response("Course not found", { status: 404 });
+    return { course: Promise.resolve(data) };
   } catch (e) {
     if (e instanceof CustomError && e.errorCode === 401) {
       throw new Response("Unauthorized", { status: 401 });
