--- src/features/courses/components/index.ts
@@ -1,2 +1,0 @@
-export * from "./CourseCreate";
-export * from "./CourseEdit";
