--- src/features/courses/css/styles.css
@@ -1,1 +1,0 @@
-@import url("./CourseForm.css");
