--- src/features/courses/types/index.ts
@@ -1,14 +1,0 @@
-export interface ICourse {
-  id: string;
-  name: string;
-  description: string;
-  startDate: string;
-  endDate: string;
-}
-
-export interface ICoursesLoader {
-  courses: Promise<ICourse[]>;
-}
-export interface ICourseLoader {
-  course: Promise<ICourse>;
-}
